pyproject.toml
[tool.poetry]
name = "alpaca"
version = "0.1.0"
description = ""
authors = ["Belg PC <tomsomoza911@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alpaca-py = "^0.33.1"
nest-asyncio = "^1.6.0"
requests = "^2.32.3"
yfinance = "^0.2.50"
joblib = "^1.4.2"
ta = "^0.11.0"
streamlit = "^1.41.1"
plotly = "^5.24.1"
scikit-learn = "^1.6.0"
typer = "^0.15.1"
asyncio = "^3.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
